Type Integer pour le BrowserRepository

+++ SquidSquAds/src/main/java/com/squidsquads/repository/stats/BrowserTypesItemRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BrowserTypesItemRepository extends JpaRepository<BrowserTypesItem, Long> {
+public interface BrowserTypesItemRepository extends JpaRepository<BrowserTypesItem, Integer> {
 
     List<BrowserTypesItem> findAllByWebsiteIDOrderByRatioDesc(Integer websiteID);
